CodeAlpha_Unemployment_Analysis/requirements.txt
pandas
numpy
matplotlib
seaborn
streamlit
Pillow